Create structs and functions for the command struct
To convert the tokenized message into a readable command, a command struct
is recommended to neatly store necessary data for it. This command
struct also contains an array of Redirect structs that stores information
on stdin/stdout redirects.

Implementation is not yet ready; deep copy implementation is still needed
and the actual parsing of the tokenized message needs to be implemented.
This will all be done in the Command struct constructor.

Before implementation, it would be helpful to create data structures for
future use. Data structures such as strings and vectors will first be
implemented before we proceed in parsing.