Add parser to tokenize words
Tokenizer returns tokens separated by whitespace and pipes/
redirects. This separation was chosen to minimize the parsing
needed to be done in the construction of the command structs
in the next step. Also, this module should handle quotes.

fgets() allows us to parse entire line at once instead of previous
scanf() command which only read the first word.