Redirection and Command structs written
Command implementation is working. Tests were written for
their constructors. The tests and the actual program
are free of memory leaks.

Some functions are a bit too long. When I write the Command
Linked List implementation, I will fix these function lengths.
For now, since we are working on pipes later, this should be
enough to unblock some.