Implement prompt and internal commands in mainloop
The prompt prints "username:cwd> " before reading in the user input into
input_str. The input string is tokenized, returning an array of words. The loop
checks if the first word is an internal command ("cd" or "exit"), and if it is,
the corresponding internal command is run.

For the actual, final version of the mainloop, the tokens will (probably) be
converted to Command structs before executing the internal or external
commands. TODO statements have been added regarding these changes.