Finish implementation of user memory access
Separate methods for checking if address is valid for read or write.
valid_read/write_addr first checks if address is not NULL and below PHYS_BASE.
Then, it attempts to read/write to the address. If there is a page fault,
instead of killing the process, we set eax to -1. get_user() and put_user()
read this value and return -1 or false if the read/write failed.