Remove MAX_ARGS
Instead, count the number of arguments and keep a static max to approximate
MAX_ARGS.

It might be cleaner to pass argc as an argument, but this would require
changing various methods to take more arguments and I don't want to modify
too much code at this point. Maybe a TODO in a future project, if it comes up
again.

args-many passes.