Fix page bug in sys_read and sys_write
Move the loading of buffer into sys_read and sys_write. This is
necessary because file_read and file_write might try to write to
a page that isn't loaded. This is because size could be greater than
the size of page. So while the first page of buffer is loaded into
memory, the next page would not be. The page fault then occurs in
the file_read or file_write call which has already acquired a lock
in devices/ide.c. Then when we evict to load this new page, we need
to write to swap which invokes ide_write and tries to reacurie the
lock in devices/ide.c which then fails.

4 page-parallel still failing. page-merge-par sometimes works.