Fix eviction of frame in munmap
We were deleting the supplemental page table entry but not the frame
in munmap which was causing a bug in the eviction when it tried to
evict that frame. The supplemental page was deleted so the frame
couldn't access the elements it was trying to access like pagedir.

page-merge-mm still hangs on bochs. All other tests pass on bochs.

page-merge-par and page-parallel fail on qemu.