Fix pinning bug
After a sector was chosen for eviction and before it was evicted, the
sector was pinned for use in another section. Add evicting bool to
account for this case. I believe this is also why VM tests are failing.
We need to prevent things from accessing a frame's page between when
it is chosen for eviction and when it is evicted.

Temporarily revert to using global lock for entire read and write
operations and removing read-ahead to get regression tests to pass.

One regression test (syn-read) is failing on bochs.
One regression test (page-merge-par) is failing on qemu.