Refactor read-ahead
Follow guidelines from Donnie's powerpoint.

Use sempahore to avoid busy waiting.

Fails syn-read and page-linear and syn-write. I think it fails syn-read
and page-linear because a different thread is the last thread to call
syn-read so a different thread releases the lock. We probably need to
refactor this to a semapahore.