Fix bug in read write lock implementation
Since any reader thread can release the global lock and it doesn't
ahve to be the reader thread that acquired the lock, we need it to be
a semaphore instead of a lock.

On bochs, syn-read and syn-write are the only regression tests failing.
On qemu, page-parallel and page-merge-*(except seq) are the only
regression tests failing.